Add utils.js for code reduction in index.js file.
